Use compiler phases instead of multi-release jar
