+++ pom.xml
@@ -183,12 +183,23 @@
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>3.11.0</version>
+                <version>3.11.1-SNAPSHOT</version>
+                <configuration>
+                    <fork>true</fork>
+                    <source>${jdkTarget}</source>
+                    <target>${jdkTarget}</target>
+                    <release>${jdkTarget}</release>
+                    <compilerArgs>
+                        <arg>-Xlint:-options</arg>
+                    </compilerArgs>
+                </configuration>
                 <executions>
                     <execution>
                         <id>default-compile</id>
                         <configuration>
-                            <release>${jdkTarget}</release>
+                            <excludes>
+                                <exclude>**/ffm/*.java</exclude>
+                            </excludes>
                         </configuration>
                     </execution>
                     <execution>
@@ -198,17 +209,16 @@
                         </goals>
                         <configuration>
                             <release>21</release>
-                            <compilerArgument>--enable-preview</compilerArgument>
-                            <compileSourceRoots>
-                                <compileSourceRoots>src/main/java21</compileSourceRoots>
-                            </compileSourceRoots>
+                            <includes>
+                                <include>**/ffm/*.java</include>
+                            </includes>
+                            <compilerArgs>
+                                <arg>--enable-preview</arg>
+                            </compilerArgs>
                         </configuration>
                     </execution>
                     <execution>
                         <id>default-testCompile</id>
-                        <configuration>
-                            <release>${jdkTarget}</release>
-                        </configuration>
                     </execution>
                 </executions>
             </plugin>
@@ -395,12 +405,12 @@
             <plugin>
                 <groupId>com.diffplug.spotless</groupId>
                 <artifactId>spotless-maven-plugin</artifactId>
-                <version>2.38.0</version>
+                <version>2.39.0</version>
                 <configuration>
                     <java>
                         <toggleOffOn />
                         <palantirJavaFormat>
-                            <version>2.35.0</version>
+                            <version>2.38.0</version>
                         </palantirJavaFormat>
                         <importOrder>
                             <order>java|javax,org,,\#</order>
